Crate substrate_wasm_builder_runner

WASM builder runner

Since cargo contains many bugs when it comes to correct dependency and feature resolution, we need this little tool. See https://github.com/rust-lang/cargo/issues/5730 for more information.

It will create a project that will call substrate-wasm-builder to prevent any dependencies from substrate-wasm-builder influencing the main project's dependencies.

For more information see https://crates.io/substrate-wasm-builder

Enums

WasmBuilderSource
The wasm-builder dependency source.

Functions

build_current_project
Build the currently built project as WASM binary.

build_current_project_with_rustflags
Build the currently built project as WASM binary and extend RUSTFLAGS with the given rustflags.